+++ CHANGES.md
@@ -34,6 +34,7 @@ Beta Releases
 * Added `AnimationViewModel.snapToTicks`, which when set to true, causes the shuttle ring on the Animation widget to snap to the defined tick values, rather than interpolate between them.
 * Added new `Grid` material.
 * Made `EllipsoidPrimitive` double-sided.
+* Fixed an error in Web Worker creation when loading Cesium.js from a different origin.
 
 ### b15 - 2013-04-01